requirements.txt
absl-py
astor
azure-cognitiveservices-vision-face
certifi
chardet
Click
cycler
Flask
gast
google-pasta
grpcio
h5py
idna
imageio
itsdangerous
Jinja2
joblib
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keras-vggface
kiwisolver
Markdown
MarkupSafe
matplotlib
mtcnn
numpy
opencv-python
pandas
Pillow
protobuf
psutil
pyparsing
python-dateutil
pytz
PyYAML
requests
scikit-learn
scipy
six
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor
torch==1.5.0+cu101
torchvision==0.6.0+cu101
tqdm
urllib3
Werkzeug
wrapt
